create custom permissions

+++ home/views.py
@@ -2,10 +2,10 @@ from rest_framework.decorators import APIView
 from rest_framework.response import Response
 from .models import Person, Question, Answer
 from .serializers import PersonSerializer,QuestionSerializer, AnswerSerializer
-from rest_framework.permissions import IsAuthenticated, IsAdminUser
+from rest_framework.permissions import IsAuthenticated
 # from rest_framework.decorators import api_view	# for fbv
 from rest_framework import status
-
+from permissions import IsOwnerOrReadOnly
 
 
 class Home(APIView):
@@ -19,14 +19,15 @@ class Home(APIView):
 		return Response(data=serialize_data.data)
 
 
-
-
-class QuestionView(APIView):
+class QuestionListView(APIView):
 	def get(self, request):
 		questions = Question.objects.all()
 		srz_data = QuestionSerializer(instance=questions, many=True).data
 		return Response(srz_data, status=status.HTTP_200_OK)
 
+
+class QuestionCreateView(APIView):
+	permission_classes = [IsAuthenticated,]
 	def post(self, request):
 		srz_data = QuestionSerializer(data=request.data)
 		if srz_data.is_valid():
@@ -34,21 +35,27 @@ class QuestionView(APIView):
 			return Response(srz_data.data, status=status.HTTP_201_CREATED)
 		return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
+class QuestionUpdateView(APIView):
+	permission_classes = [IsOwnerOrReadOnly,]
+
 	def put(self, request, pk):
 		question = Question.objects.get(pk=pk)
+		self.check_object_permissions(request, question)
 		srz_data = QuestionSerializer(instance=question, data=request.data, partial=True)
 		if srz_data.is_valid():
 			srz_data.save()
 			return Response(srz_data.data, status=status.HTTP_200_OK)
 		return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
+class QuestionDeleteView(APIView):
+	permission_classes = [IsOwnerOrReadOnly,]
+
 	def delete(self, request, pk):
 		question = Question.objects.get(pk=pk)
+		self.check_object_permissions(request, question)
 		question.delete()
-		return Response({'message': 'question deleted'}, status=status.HTTP_200_OK)
-
-
-
